requirements.txt
Django==4.2
psycopg2-binary==2.9.9
Pillow==9.4.0
# django-environ==0.11.2 
djangorestframework==3.14.0
django-cors-headers==4.4.0 
django-filter==24.3 